Fix mat_lines.c
The code was avoiding completely any type of sync. Therefore for big
samples you could see that the output wasn't sorted.

In order to fix I waited for the child process and outputed it's result.
Now the sequence was decreasing so I had to start from MAX_LINE - 1 to 0

Furthermore I was creating a process on the last cycle. So fixed that
too.